+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### 🚀 Features
 
+- Inject secrets via 1password
+- Inject secrets via bitwarden
 - Add certbot_email configuration for nginx
 
 ### 🚜 Refactor